pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["wassname <git@wassname.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
torch = {version = "^2.1.0+cu118", source = "pytorch"}
matplotlib = "^3.8.0"
loguru = "^0.7.2"
einops = "^0.3.1"
torchinfo = "^1.8.0"
accelerate = "^0.24.1"
peft = "^0.5.0"
bitsandbytes = {url = "https://github.com/TimDettmers/bitsandbytes/releases/download/0.41.0/bitsandbytes-0.41.0-py3-none-any.whl"}
transformers = "4.34.0"
tqdm = "^4.66.1"
wandb = "^0.12.6"
ale-py = "^0.7.4"
pygame = "^2.5.2"
psutil = "^5.9.6"
protobuf = "^3.10.0"
opencv-python = "^4.8.1.78"
hydra-core = "^1.3.2"
torchvision = "^0.16.0"
numpy = ">=1.18.0"
gym = {version = "0.22.0", extras = ["accept-rom-license", "atari"]}
scipy = "^1.11.3"
crafter = "^1.8.2"
minihack = "^0.1.5"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"
ruff = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
